+++ lib/type_ahead.dart
@@ -10,11 +10,13 @@ class TypeAhead<T> extends StatelessWidget {
   TypeAhead(this.api) : search = new Search(api);
 
   Widget buildState(ServiceState serviceState) {
-    if(serviceState.isProcessing)
-    {
-      return new Center(child: new Container( margin: const EdgeInsets.all(10.0), child: new CircularProgressIndicator()));
-    }
-    else if (serviceState.result?.model != null && serviceState.result.model.length > 0) {
+    if (serviceState.isProcessing) {
+      return new Center(
+          child: new Container(
+              margin: const EdgeInsets.all(10.0),
+              child: new CircularProgressIndicator()));
+    } else if (serviceState.result?.model != null &&
+        serviceState.result.model.length > 0) {
       ListView lb = new ListView.builder(
           shrinkWrap: true,
           itemCount: serviceState.result?.model?.length ?? 0,
@@ -43,16 +45,11 @@ class TypeAhead<T> extends StatelessWidget {
           final serviceState = snapshot.data;
 
           return new Column(children: [
-            new TextField(                
-                onChanged: (String s) => search.onTextChanged.add(s)),
+            new TextField(onChanged: (String s) => search.onTextChanged.add(s)),
             new Row(
-              children: <Widget>[
-                new Expanded(child: buildState(serviceState))
-              ],
+              children: <Widget>[new Expanded(child: buildState(serviceState))],
             )
           ]);
         });
   }
-  
 }
-