+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+Initial version.  A very simple looking, but easy to use TypeAhead widget
+
+## 0.0.2
+
+Upgrading to flutter 0.6.0.